move module chat screen
