+++ app/src/main/java/com/jin/honey/feature/navigation/RootNavigation.kt
@@ -42,8 +42,8 @@ import com.jin.domain.usecase.RemoveIngredientInCartItemUseCase
 import com.jin.domain.usecase.RemoveMenuInCartUseCase
 import com.jin.ui.category.CategoryScreen
 import com.jin.ui.category.CategoryViewModel
-import com.jin.honey.feature.chat.ui.ChatScreen
-import com.jin.honey.feature.chat.ui.ChatViewModel
+import com.jin.ui.chat.ChatScreen
+import com.jin.ui.chat.ChatViewModel
 import com.jin.domain.favorite.FavoriteRepository
 import com.jin.domain.usecase.GetFavoriteMenuUseCase
 import com.jin.domain.usecase.GetRecentlyMenuUseCase
